create new product and upload image to cloudinary

+++ pages/api/products/[id].jsx
@@ -2,8 +2,8 @@ import dbConnect from "../../../lib/dbConnect";
 import Product from "../../../models/Product";
 
 export default async function handler(req, res) {
-  const { method, query } = req;
-    
+  const { method, query, cookies } = req;
+  const token = cookies.token;
   dbConnect();
 
   if (method === "GET") {
@@ -20,6 +20,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "POST") {
+    if(!token || token !== process.env.token){
+      return res.status(401).json("Not authenticated!")
+    }
     try {
       return res.status(201).json({
         status: "success",
@@ -31,8 +34,10 @@ export default async function handler(req, res) {
   }
   
   if (method === "DELETE") {
+    if(!token || token !== process.env.token){
+      return res.status(401).json("Not authenticated!")
+    }
     try {
-
       await Product.findOneAndDelete(query.id);
 
       return res.status(204).json({
